add sample_rate to hotfilm voltage attributes
check for sample_rate and break up blocks when the rate changes, then store
the rate in the netcdf as an attribute of the hotfilm voltage variables.  also
fix integer attributes to use np.int32 type, so they are encoded in netcdf as
an int type instead of long long (LL).

+++ hotfilm/hotfilm_wind_speed_dataset.py
@@ -29,8 +29,8 @@ class HotfilmWindSpeedDataset:
         Return a time coordinate for the calibration period.
         """
         attrs = {'long_name': 'Calibration period begin time',
-                 'period_seconds': hfc.period_seconds,
-                 'mean_interval_seconds': hfc.mean_interval_seconds}
+                 'period_seconds': np.int32(hfc.period_seconds),
+                 'mean_interval_seconds': np.int32(hfc.mean_interval_seconds)}
         timed = xr.DataArray(name=self.CALIBRATION_TIME,
                              data=[hfc.begin],
                              dims=[self.CALIBRATION_TIME],
